+++ src/components/Capture/Document.tsx
@@ -193,7 +193,7 @@ const Document = (props: Props) => {
     side,
     trackScreen,
     uploadFallback = true,
-    useLiveDocumentCapture,
+    useLiveDocumentCapture = true,
     useWebcam,
   } = props